move all for tests to conftest
